requirements.txt
pysimplegui==4.70.1
paho-mqtt==1.6.1
